+++ carp_audio_package/lib/audio_probe.dart
@@ -6,13 +6,119 @@
  */
 part of audio;
 
-// TODO - PERMISSIONS
-// This probe needs PERMISSIONS to use
-//    - audio recording / microphone
-//    - file access / storage
-//
-// If these permissions are not set, the app crashes....
-// See issue on github.
+/// A probe recording audio from the microphone. It starts recording on [resume()]
+/// and stops recording on [pause()], post its [AudioDatum] to the [events] stream.
+/// Use a [AudioMeasure] to configure this probe.
+///
+/// This probe can be used in a [PeriodicTrigger], which allows for periodic sampling
+/// of audio by specifying the [frequency] and [duration].
+/// It is important that the recording duration is not longer than the sampling frequency,
+/// i.e. this probe does **not** allow for overlapping recordings.
+///
+/// Note that this probe generates a lot of data and should be used with caution.
+///
+/// Also note that this probe records raw sound directly from the microphone and hence
+/// records everything - including human speech - in its proximity.
+///
+/// The Audio probe generates an [AudioDatum] that holds the meta-data for each recording
+/// along with the actual recording in an audio file. How to upload this data to a data backend
+/// is up to the implementation of the [DataManager], which is used in the [Study].
+class AudioProbe extends DatumProbe {
+  static const String AUDIO_FILE_PATH = 'audio';
+
+  String studyId;
+  String soundFileName;
+  String _path;
+  bool _isRecording = false;
+  DateTime _startRecordingTime, _endRecordingTime;
+  FlutterSound _flutterSound = new FlutterSound();
+
+  Future<void> onInitialize(Measure measure) async {
+    assert(measure is AudioMeasure);
+    super.onInitialize(measure);
+    this.studyId = (measure as AudioMeasure).studyId;
+  }
+
+  Future<void> onStart() async {} // do nothing on start -- only record on resume()
+
+  Future<void> onRestart() async {
+    super.onRestart();
+    this.studyId = (measure as AudioMeasure).studyId;
+  }
+
+  Future<void> onResume() async {
+    soundFileName = await _startAudioRecording().catchError((err) => controller.addError(err));
+  }
+
+  Future<void> onPause() async {
+    // when pausing the audio sampling, stop recording and collect the datum
+    await _stopAudioRecording().catchError((err) => controller.addError(err));
+    getDatum().then((Datum data) {
+      if (data != null) controller.add(data);
+    }).catchError((error, stacktrace) => controller.addError(error, stacktrace));
+  }
+
+  Future<void> onStop() async {
+    if (_isRecording) await onPause();
+    _flutterSound.stopRecorder();
+    _flutterSound = null;
+    super.onStop();
+  }
+
+  Future<Datum> getDatum() async {
+    return (soundFileName != null)
+        ? AudioDatum(
+            filename: soundFileName, startRecordingTime: _startRecordingTime, endRecordingTime: _endRecordingTime)
+        : null;
+  }
+
+  Future<String> _startAudioRecording() async {
+    if (_isRecording) throw new Exception('AudioProbe is already running');
+    _startRecordingTime = DateTime.now();
+    _isRecording = true;
+
+    return _flutterSound.startRecorder(filename);
+  }
+
+  Future<void> _stopAudioRecording() async {
+    _endRecordingTime = DateTime.now();
+    _isRecording = false;
+    await _flutterSound.stopRecorder();
+  }
+
+  /// Returns the local path on the device where sound files can be stored.
+  /// Creates the directory, if not existing.
+  Future<String> get path async {
+    if (_path == null) {
+      // get local working directory
+      final localApplicationDir = await getApplicationDocumentsDirectory();
+      // create a sub-directory for sound files
+      final directory =
+          await Directory('${localApplicationDir.path}/${FileDataManager.CARP_FILE_PATH}/$studyId/$AUDIO_FILE_PATH')
+              .create(recursive: true);
+
+      _path = directory.path;
+    }
+    return _path;
+  }
+
+  /// Returns the  filename of the sound file.
+  /// The filename format is
+  ///    * Android : `audio-yyyy-mm-dd-hh-mm-ss-ms.mp4`
+  ///    * iOS : `audio-yyyy-mm-dd-hh-mm-ss-ms.m4a`
+  String get filename {
+    String created =
+        DateTime.now().toString().replaceAll(" ", "-").replaceAll(":", "-").replaceAll("_", "-").replaceAll(".", "-");
+    String type = Platform.isIOS ? 'm4a' : 'mp4';
+    return 'audio-$created.$type';
+  }
+
+  /// Returns the full file path to the sound file.
+  Future<String> get filePath async {
+    String dir = await path;
+    return "$dir/filename";
+  }
+}
 
 /// A probe recording audio from the microphone.
 ///
@@ -28,7 +134,7 @@ part of audio;
 /// The Audio probe generates an [AudioDatum] that holds the meta-data for each recording
 /// along with the actual recording in an audio file. How to upload this data to a data backend
 /// is up to the implementation of the [DataManager], which is used in the [Study].
-class AudioProbe extends BufferingPeriodicProbe {
+class DeprecatedAudioProbe extends BufferingPeriodicProbe {
   static const String AUDIO_FILE_PATH = 'audio';
 
   String studyId;