+++ carp_audio_package/CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.6.1
 * allignment with `carp_mobile_sensing` v. 0.6.1.
+* improvement to the audio probe based.
 
 ## 0.6.0
 * update to `carp_mobile_sensing` version 0.6.0